+++ packages/themes/CHANGELOG.md
@@ -2,3 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# 1.1.0 (2022-06-15)
+
+
+### Features
+
+* **button:** button component first implementation ([#28](https://github.com/hey-car/heycar-uikit/issues/28)) ([8b267fb](https://github.com/hey-car/heycar-uikit/commit/8b267fb5cf93cac5cab60889bba036b541349785))