+++ packages/vars/CHANGELOG.md
@@ -2,3 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# 1.1.0 (2022-06-15)
+
+
+### Bug Fixes
+
+* add missing status colors ([#9](https://github.com/hey-car/heycar-uikit/issues/9)) ([4bbc396](https://github.com/hey-car/heycar-uikit/commit/4bbc39660a28eb36c92b1378dfe1be13c0346be9))
+
+
+### Features
+
+* **button:** button component first implementation ([#28](https://github.com/hey-car/heycar-uikit/issues/28)) ([8b267fb](https://github.com/hey-car/heycar-uikit/commit/8b267fb5cf93cac5cab60889bba036b541349785))
+* **typography:** typography component ([#34](https://github.com/hey-car/heycar-uikit/issues/34)) ([48aa261](https://github.com/hey-car/heycar-uikit/commit/48aa26142d3a0406c732bbe7e8624494172828b7))